added landing page, chooseGameMode and gameSetup - todo: player name as input field
