+++ package.json
@@ -12,10 +12,10 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.3",
     "react-stomp": "^5.1.0",
     "sockjs": "^0.3.24",
-    "sockjs-client": "^1.6.1",
-    "stompjs": "^2.3.3"
+    "sockjs-client": "^1.6.1"
   },
   "devDependencies": {
     "@types/react": "^18.2.56",
@@ -23,11 +23,14 @@
     "@types/sockjs-client": "^1.5.4",
     "@types/stompjs": "^2.3.9",
     "@vitejs/plugin-react": "^4.2.1",
+    "autoprefixer": "^10.4.19",
     "eslint": "^8.56.0",
     "eslint-plugin-react-hooks": "^4.6.0",
+    "postcss": "^8.4.38",
     "prettier": "3.2.5",
     "sockjs-client": "^1.6.1",
     "stompjs": "^2.3.3",
+    "tailwindcss": "^3.4.3",
     "typescript": "^5.3.3",
     "typescript-eslint": "^7.0.1",
     "vite": "^5.1.4",
