+++ src/MapDisplay.css
@@ -1,7 +1,5 @@
 .MapDisplay-map-container {
-
     border: 3px solid #000000;
-
 }
 
 .MapDisplay-row {
