+++ src/at.fhv.GameSetup/Lobby.tsx
@@ -1,0 +1,3 @@
+export default function Lobby() {
+    
+}
