+++ src/index.tsx
@@ -1,8 +1,13 @@
-import ReactDOM from 'react-dom/client';
-import App from './App';
-import './index.css';
+import ReactDOM from "react-dom/client";
+import App from "./App";
+import "./index.css";
+import { BrowserRouter } from "react-router-dom";
+import React from "react";
 
-const rootNode = document.getElementById('root') as HTMLElement;
-const root = ReactDOM.createRoot(rootNode);
-
-root.render(<App />);
+ReactDOM.createRoot(document.getElementById("root")!).render(
+  <React.StrictMode>
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  </React.StrictMode>,
+);
